refac: 커스텀 훅, 유틸함수 추가
